Refactor SMT pad drawing: integrate drawPadShape into createCopperLayerTexture and remove pad-drawer module
